added more ros subscirbers and publishers

+++ ArmStuff/Armduino/Carousel.cpp
@@ -2,7 +2,7 @@
 #include "Motor.h"
 #include "Switch.h"
 #include "Carousel.h"
-
+#include "ros.h"
 Motor rotationMotor;
 Motor crankMotor;
 Switch closedSwitch;
@@ -26,19 +26,19 @@ void Carousel::Close(){
 	_Direction = ((closedSwitch.getState() == false) ? -1 : 0);
 	crankMotor.doMotor(_Direction);
 }
-void Carousel::rotate(int _index){
+void Carousel::rotate(uint16_t _index){
 	int goTo = ((_index + index > 5)?index-_index:index+_index);
 	while(index < _index){
 		index = ((indexSwitch.getState() == HIGH)? index++ : index);
 		rotationMotor.doMotor(1);
 	}
 }
-void Carousel::fillIndex(int _index){
+void Carousel::fillIndex(uint16_t _index){
 	fullIndexes[_index] = true;
 }
-void Carousel::emptyIndex(int _index){
+void Carousel::emptyIndex(uint16_t _index){
 	fullIndexes[_index] = false;
 }
-bool Carousel::checkIndex(int _index){
+bool Carousel::checkIndex(uint16_t _index){
 	return fullIndexes[_index];
 }
